+++ src/System/Http/Request.php
@@ -119,7 +119,6 @@ class Request implements \ArrayAccess, \IteratorAggregate
      * @param array<string, string> $cookies
      * @param array<string, string> $files
      * @param array<string, string> $headers
-     * @param ?string               $rawBody
      */
     public function __construct(
         string $url,
@@ -131,7 +130,7 @@ class Request implements \ArrayAccess, \IteratorAggregate
         array $headers = [],
         string $method = 'GET',
         string $remoteAddress = '::1',
-        ?string $rawBody = null
+        ?string $rawBody = null,
     ) {
         $this->initialize($url, $query, $post, $attributes, $cookies, $files, $headers, $method, $remoteAddress, $rawBody);
     }
@@ -145,7 +144,6 @@ class Request implements \ArrayAccess, \IteratorAggregate
      * @param array<string, string> $cookies
      * @param array<string, string> $files
      * @param array<string, string> $headers
-     * @param ?string               $rawBody
      *
      * @return self
      */
@@ -159,7 +157,7 @@ class Request implements \ArrayAccess, \IteratorAggregate
         array $headers = [],
         string $method = 'GET',
         string $remoteAddress = '::1',
-        ?string $rawBody = null
+        ?string $rawBody = null,
     ) {
         $this->url             = $url;
         $this->query           = new Collection($query);
@@ -193,7 +191,7 @@ class Request implements \ArrayAccess, \IteratorAggregate
         ?array $attributes = null,
         ?array $cookies = null,
         ?array $files = null,
-        ?array $headers = null
+        ?array $headers = null,
     ) {
         $dupplicate = clone $this;